Fix: typos in clone commands were creating empty repos
Changes to the logic in serve() made it possible for code flow to fall
through further than initially expected. ONLY attempt to init a new
repo if we are trying to WRITE to a non-existent repo. That way,
trying to read from a non-existent repo will fail as it should.